[추가] Data 및 Domain 레이어 분리
