+++ KuringApp/KuringApp/ContentView.swift
@@ -2,7 +2,6 @@
 // Copyright (c) 2024 쿠링
 // See the 'License.txt' file for licensing information.
 //
-
 import SwiftUI
 import ColorSet
 import CampusUI
