+++ KuringApp/KuringWidget/SubscriptionWidget/ViewModel/SubscriptionWidgetAppIntent.swift
@@ -30,7 +30,8 @@ struct SubscriptionWidgetAppIntent: AppIntent {
     init(selection: String) {
         self.selection = selection
     }
-    
+   
+    @MainActor
     func perform() async throws -> some IntentResult {
         var subscriptions = DataStorageManager.shared.subscriptions
         
@@ -46,54 +47,4 @@ struct SubscriptionWidgetAppIntent: AppIntent {
         
         return .result()
     }
-}
-
-/// 공지 구독
-struct SubscriontionRepository {
-    @Dependency(\.kuringLink) var kuringLink
-    
-    /// 대학 구독 정보 갱신
-    ///
-    /// - Note: Swift 6.0이후 throw에 Error 타입 지정할 수 있음.
-    /// - Note: `Result`로 처리하면 더 명확하나, 기존 설계가 `Bool`로 응답 결과를 결정하고 있어서 해당 흐름을 따라가도록 설계
-    func updateUnivSubscription(selections: Set<NoticeProvider>) async -> Bool {
-        let typeNames = DataStorageManager.shared.subscriptions.compactMap { $0.name }
-        
-        let result = try? await kuringLink.subscribeUnivNotices(typeNames)
-        return result ?? false
-    }
-}
-
-struct WidgetSubscriponProvider {
-    @Dependency(\.kuringLink) var kuringLink
-    @Dependency(\.subscriptions) var subscriptions
-    
-    private let subscriontionRepository = SubscriontionRepository()
-    
-    /// 위젯 구독 선택에 따른 상태 변경
-    func selection(noticeProvider: NoticeProvider) async {
-        var snapshot = DataStorageManager.shared.subscriptions
-        
-        if snapshot.contains(where: { $0 == noticeProvider }) {
-            // 기존에 구독했던 경우 > 삭제
-            snapshot.remove(noticeProvider)
-        } else {
-            // 구독하지 않았던 경우 > 등록
-            snapshot.insert(noticeProvider)
-        }
-        
-        // 서버 데이터 통신
-        let result = await subscriontionRepository.updateUnivSubscription(
-            selections: snapshot
-        )
-        
-        // 로컬 데이터 갱신
-        switch result {
-        case true:
-            subscriptions.update(snapshot)
-        case false:
-            break
-        }
-    }
-    
 }
